Remove Bufferline plugin.
Bufferline is trying to bring the VSCode mindset to vim, which does not work.
Vim does not have a concept of "Tabs" in the same way that other editors have.
"Tabs" in vim are window layouts.
It's not helpful to try to force the concept of "vim buffer" into "IDE tabs"
as you can't assign a buffer to a window or tab and you can't change the order of buffers.

It's much better to just let go of the concept of "each buffer is a tab" and use
VIM native navigation method like "Go To Definition", "Find Usage", etc. instead.

If you really need to work with a set of unrelated files, just use a plugin like Harpoon instead.