Buf fix in lazy loading
Defnining the same keys for lazy loading as for the action will result
in mappings becomming overriten/deleted.

Lazy.nvim will automatically load the plugins as soon as they're
"require"ed, so there is no need for defining keys for plugins that
we've defined key bindings for.

Also, for lazy loading to work correctly, we need to "require" the
plugin just when we actually need it, so we need to move all "require"s
inside the key-bindings to files.